+++ class/Tweakr.php
@@ -92,6 +92,12 @@ class Tweakr{
             Tweakr\API::restrictRestApiAccess();
         }
 
+        // monitoring
+        if ($this->_settingsManager->getOption('monitoring-enabled')){
+            $monitoring = new Tweakr\Monitoring();
+            $monitoring->registerRestEndpoint();
+        }
+
         // New User Notifications
         // ------------------------------------------------------------------
         Tweakr\UserNotification::processNewRegistrations($this->_settingsManager->getOption('user-registration-email-notification'));
@@ -363,14 +369,14 @@ class Tweakr{
             // plugin installed ?
             if ($version == '0.0.0'){
                 // store new version
-                update_option('tweakr-version', '1.4-BETA2');
+                update_option('tweakr-version', '1.4-BETA3');
 
             // plugin upgraded ?
-            }else if (version_compare('1.4-BETA2', $version, '>')){
+            }else if (version_compare('1.4-BETA3', $version, '>')){
                 // run upgrade hook
                 if ($i->_wp_plugin_upgrade($version)){
                     // store new version
-                    update_option('tweakr-version', '1.4-BETA2');
+                    update_option('tweakr-version', '1.4-BETA3');
 
                     // set flag (string!)
                     update_option('tweakr-upgrade', 'true');